Fixes #25 by checking for exact platform matches first and including non-native matches when only prereleases are installed. (#27)

+++ src/manage/install_command.py
@@ -57,7 +57,7 @@ def _expand_versions_by_tag(versions):
                 yield {**v, "tag": t}
 
 
-def select_package(cmd, source, tag, cache, *, urlopen=_urlopen, by_id=False):
+def select_package(index_downloader, tag, platform=None, *, urlopen=_urlopen, by_id=False):
     """Finds suitable package from index.json that looks like:
     {"versions": [
       {"id": ..., "company": ..., "tag": ..., "url": ..., "hash": {"sha256": hexdigest}},
@@ -67,7 +67,7 @@ def select_package(cmd, source, tag, cache, *, urlopen=_urlopen, by_id=False):
     """
 
     first_exc = None
-    for index in IndexDownloader(cmd.source, Index, {}, cache):
+    for index in index_downloader:
         try:
             if by_id:
                 for v in index.versions:
@@ -76,6 +76,11 @@ def select_package(cmd, source, tag, cache, *, urlopen=_urlopen, by_id=False):
                 raise LookupError("Could not find a runtime matching '{}' at '{}'".format(
                     tag, sanitise_url(index.source_url)
                 ))
+            if platform:
+                try:
+                    return index.find_to_install(tag + platform)
+                except LookupError:
+                    pass
             return index.find_to_install(tag)
         except LookupError as ex:
             first_exc = ex
@@ -334,6 +339,10 @@ def print_cli_shortcuts(cmd):
             LOGGER.info("Installed %s to %s", i["display-name"], i["prefix"])
 
 
+def _same_install(i, j):
+    return i["id"] == j["id"] and i["sort-version"] == j["sort-version"]
+
+
 def _find_one(cmd, source, tag, *, installed=None, by_id=False):
     if by_id:
         LOGGER.debug("Searching for Python with ID %s", tag)
@@ -341,7 +350,9 @@ def _find_one(cmd, source, tag, *, installed=None, by_id=False):
         LOGGER.verbose("Searching for Python matching %s", tag)
     else:
         LOGGER.verbose("Searching for default Python version")
-    install = select_package(cmd, source, tag, DOWNLOAD_CACHE, by_id=by_id)
+
+    downloader = IndexDownloader(source, Index, {}, DOWNLOAD_CACHE)
+    install = select_package(downloader, tag, cmd.default_platform, by_id=by_id)
 
     if by_id:
         return install
@@ -365,7 +376,8 @@ def _find_one(cmd, source, tag, *, installed=None, by_id=False):
 
     # Return the package if it was requested in a way that wouldn't have
     # selected the existing package (e.g. full version match)
-    if not install_matches_any(existing[0], [tag]):
+    if (not _same_install(install, existing[0])
+        and not install_matches_any(existing[0], [tag])):
         if cmd.ask_yn("!Y!Your existing %s install will be replaced by " +
                       "%s. Continue?!W!", existing[0]["display-name"],
                       install["display-name"]):
@@ -530,6 +542,7 @@ def execute(cmd):
         cmd.tags = []
         for arg in cmd.args:
             if arg.casefold() == "default".casefold():
+                LOGGER.debug("Replacing 'default' with '%s'", cmd.default_tag)
                 cmd.tags.append(tag_or_range(cmd.default_tag))
             else:
                 try:
@@ -666,6 +679,9 @@ def execute(cmd):
                 except LookupError:
                     LOGGER.error("Failed to find a suitable install for '%s'.", tag)
                     raise NoInstallFoundError()
+                except (AssertionError, AttributeError, TypeError):
+                    # These errors should never happen.
+                    raise
                 except Exception as ex:
                     LOGGER.debug("Capturing error in case fallbacks fail", exc_info=True)
                     first_exc = first_exc or ex
